Added authorization and file sharing backend

+++ FileSharingWebsite/Controllers/FilesController.cs
@@ -1,10 +1,12 @@
 ﻿using System;
 using System.Security.Claims;
 using System.Threading.Tasks;
+using FileSharingWebsite.Authorization;
 using FileSharingWebsite.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Razor.Language.CodeGeneration;
 using Microsoft.Extensions.Logging;
 using Microsoft.Net.Http.Headers;
 
@@ -15,15 +17,18 @@ namespace FileSharingWebsite.Controllers
     {
         private readonly IFileService _fileService;
         private readonly ILogger<FilesController> _logger;
-        public FilesController(IFileService fileService, ILogger<FilesController> logger)
+        private readonly IAuthorizationService _authorization;
+
+        public FilesController(IFileService fileService, ILogger<FilesController> logger, IAuthorizationService authorization)
         {
             _fileService = fileService;
             _logger = logger;
+            _authorization = authorization;
         }
         // GET: File
         public async Task<ActionResult> Index()
         {
-            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var userId = User.FindFirstValue(ClaimTypes.Name);
             var files = await _fileService.UserFilesAsync(userId);
             return View(files);
         }
@@ -42,8 +47,14 @@ namespace FileSharingWebsite.Controllers
 
         public async Task<ActionResult> Download(Guid id)
         {
-
             var fileDetails = await _fileService.GetFileDetailsAsync(id);
+            var authResult = await _authorization.AuthorizeAsync(User, fileDetails, "AccessSharedFil");
+
+            if (!authResult.Succeeded)
+            {
+                return new ForbidResult();
+            }
+
             var stream = await _fileService.GetFileStreamAsync(id);
             await _fileService.IncrementFileDownloadsAsync(id);
             _logger.LogInformation($"Starting download of file {id}");
@@ -58,7 +69,7 @@ namespace FileSharingWebsite.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Create(IFormFile file)
         {
-            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var userId = User.FindFirstValue(ClaimTypes.Name);
             var savedFile = await _fileService.SaveAsync(file, userId);
 
             _logger.LogInformation($"User {userId} saved new file {savedFile.Id}");
@@ -70,6 +81,14 @@ namespace FileSharingWebsite.Controllers
         [HttpGet]
         public async Task<ActionResult> Delete(Guid id, IFormCollection collection)
         {
+            // make sure this file belongs to current user
+            var file = await _fileService.GetFileDetailsAsync(id);
+            var authResult = await _authorization.AuthorizeAsync(User, file, "ManageFile");
+            if (!authResult.Succeeded)
+            {
+                return new ForbidResult();
+            }
+
             try
             {
                 await _fileService.DeleteFileAsync(id);
